Fix #480 in LambdaLift
Fix two errors, where the first masked the second:

1) Variables defined in a method are not free variables of that method. This was mispredicated before
and caused liftedOwner to be updated to the class enclosing the method, thereby preventing any method
that refers to a local parameter or symbol to be hoisted.

Once this was fixed, methods were hoisted too far out. Test case in #480a, which takes code from Definitions.
This was fixed by

2) markFree is updated to do an additional narrowLiftedOwner so that, if a free variable gets a proxy
in an enclosing class, any inner classes and methods are kept within that class.