Drop restriction that SuperAccessors should not touch patterns
Not sure why we need to do this, and in any case it's not sure
what constitutes a pattern. There are certainly some parts of
patterns (e.g. prefixes of unapplies, or their implicit arguments)
that should be transformed under SuperAccessors, so the previous
condition was too coarse.

We include the test case that motivated the restriction. It looks like
it works now.