Make computeDenot take overrides into account
If previous denotation was a sym denotation, it might be overridden
by a different symbol in a new phase or run. So it is not correct to
simply return the current version of the symbol, as was done before.

We now recompute the member if there is a chance that the symbol could
be overridden.