Dropped ??? in arrayclass
The patch caused the built dotty not to run anymore. Since the prupose of the ???
was to find out where it was called I inserted the stacktrace that caused dotc
to fail.

I wonder how this passed the tests? Dotc became completely useless for me when I
ran it from the command line.