Avoid cycle between findMember and &
As test case IterableDotty shows, there can be a cycle

  findMember  -->  &  -->  <:<  -->  hasMatchingMember -->  findMember

which leads to an infinite recursion. This commit avoids this by
keeping track of names of outstanding findMember searches and falling
back to forming AndTypes instead of `&' is we repeat a findMember
search on a name seen before after a certain recursion depth.