Introduce harmonization of numeric arguments
Harmonization is Dotty's alternative to Scala 2's notion of weak conformance. It is less powerful but also
less entangled with the core type system. The idea is that in some specific contexts trees that
all have primitive numeric types will be converted as necessary so that they all have the same numeric type.
These tree sets are:

 - the two branches of an if
 - the alternatives of a match
 - the body together with the catch blocks of a try
 - the arguments of a vararg parameter

Examples are in the test file, harmonize.scala.