Make LazyVals respect scope hygiene.
The previius commit on "Refine & for Denotations" caused
NotDefinedHere errors in the backend when compiling dotc/typer.
These were tracked down to three occurrences in LazyVals where `enter` instead
of `enteredAfter` was used. `enter` will enter a symbol in an unknown set of
previous phases. Transformations that traverse scope (like erasedMembers in
TypeErasure will then see the denotations of these symbols outside the
scope in which they are defined.