Better error message for RefChecks
Trying to diagnose failure of t2503 and t5577 which were moved
to disabled in the time the tests did not work. These failed
locally for me, but the checkin tests succeeded (since the checkin
tests tested something else).

Added better diagnostics to RefChecks which now mention candidate
implementations of missing abstract members that have the right name but
not the right type.