Avoid double context creation when modes change on a fresh context.
Move addMode and friends to two decorators, one for Context, the other
for FreshContext. Implement behavior accordingly.

This avoids creating two contexts in situations like:

     c.fresh.setxploreTyperState.addMode(...)

Mow we can write

     c.fresh.addMode(...).setExploreTyperState

Because addMode returns a fresh context when applied to a fresh context.
Note that we specifically do not want virtual dispatch of addMode, that's
why it was moved to a decorator.

Also: removed mention of ".fresh: when just forllowed by an addMode, because
that one is redundant.