Stratify classTags in DottyPredef
Test case in pending/run/array-addition.scala. The problem was that
implicit search failed to insert `genericArrayOps` around `Array()`.
The reason was that the implicit parameter for `genericArrayOps` was
ambiguous, multiple `classTag` methods matched the expected type
`ClassTag[?T >: Nothing]`.

Stratifying ClassTags in DottyPredef into layers according
to implicit priority avoids this problem. This is a temporary solution
until we have proper ClassTag handling in place.