Fix size update on `mutable.TreeMap#clear()`
The previous implementation has a major bug - although `clear()` sets the root node to `null`, the `size` attribute of the `Tree` was not updated. This effectively meant that even after a `map.clear()`, a call to `map.size` would still yield the old size of the map.

The scalacheck test suite was updated to contemplate this issue.