Drop TraitConstructors phase
It did not do enough to carry its own weight, in particular because DenotationTransformers do have
a price - every encountered denotation in the whole program is passed through them. The name change
from <init> to $init$ was all it did, that is now rolled into Mixin.

Also renamed IMPLCLASS_CONSTRUCTOR to TRAIT_CONSTRUCTOR.