Make compareHK work if other side is already a lambda.
There ws a missing case, where a lambda projection `T#Apply` is compared
to something to a type lambda. In that case we should compare
the prefix `T` directly wiht the type lambda.

With that modification i553 works if the canBuildFrom implicit is imported.
But the implicit is not yet found in the type scope.