Refine lookupRefined for Wildcard prefixes
lookupRefined used to fall back to unbounded Wildcard, given a Wildcard prefix.
But that loses bounds when called from wildApprox, which in turn loses companion
objects in the implicit scope of a type.

In `i553-shuffle.scala` this happened because we projected with #Apply on a type that
approximated to a bounded wildcard.

The fix keeps the bounds, projecting in turn on them.