Another solution to the problem of discarded types in wildApprox
lookupRefined on a Wildcard prefix will return an unbounded Wildcard type.
This can destroy information that's important for determining the companions
making up the implicit scope of a type. We handle this case by maintaining
a collection for types that were discarded in that way and taking these
types intoa ccount separately when computing companions.

Examples where this is necessary include `i553-shuffle.scala` abd `vector1.scala`.