Ignore implicit parameters when interpolating type variables.
Fixes #738.

The previous scheme interpolated type variables according to the
full variance of the result type. This means that the interpolation direction
is downwards if a type varaible occurs like this:

   (implicit x: T)T

but it is upwards if it occurs like this

   (implicit x: T)Unit

For a normal method this makes sense, since the upper bound of T gives
you the most general function. But if the method is implicit, it makes
no sense, since upward instantiation means that we likely get ambiguity
errors.

Interestingly the fix causes a different problem (exemplified by failures
in run/stream-stack-overflow.scala, and also in dotc itself), which will
be fixed in the next commit.