Instantiate all type vars occurring in formal implicit parameters downwards.
With the previous commit, we had an interesting failure in relation to CanBuildFrom.

Say we have

  expr: (implicit cbf: CanBuildFrom[Repr, Elem, That]): That

for some type vars Repr >: R, Elem >: R, That. CanBuildFrom is defined as

  trait CanBuildFrom[-Repr, -Elem, +That]

Since Repr and Elem appear doubly contravariantly - that is covariantly - in

   (implicit cbf: CanBuildFrom[Repr, Elem, That]): That

it means they are minimized. With the change of the last commit, implicit
parameters are ignored for interpolation, so Repr, and Elem are not interpolated.
But they will be instantiated in `adaptNoArgs` as type variables occurring in an
implicit type. Only now the instantiation is according to the variance of the
argument, so the type variables are maximized to Any!

The same problem would have occurred earlier, if `Repr` or `Elem` would be part
of the result type of a method, because then they would also have to be maxmimized.
But by soem coincidence that code never was seen in the wild in the collection framework.

The fix minimizes all type variables in formal implicit parameters, independently
of their variance in the parameter type. This follows a suggestion of @extempore
a while ago. There's another half to it that still needs to be implemented: when
comparing multiple implicits for which one is best, we also need to ignore variance.