Expand names of super accessors
Scalac seems to treat superaccessors of traits in a rather peculiar way. They are
left unexpanded so that a class implementing several traits with the same superaccessors
will get duplicate methods with the same name. It seems this is then resolved in the backend.

Here we solve the issue by expanding the names of trait super accessors immediately.