SI-9407  Vector implementation bit-shift bugfix
Fixed logically incorrect or unnecessary code in Vector as reported by Dirk Toewe.

No tests. Because of the size of the vectors, tests would be impractically slow.  Also, the logic is quite clear: when you are recursing through a tree, using the wrong bit shift means you hit the wrong part of the tree, and when you create and then always overwrite a mutable var, you should just not do it to begin with.