SI-9388  Fix Range behavior around Int.MaxValue
terminalElement (the element _after_ the last one!) was used to terminate foreach loops and sums of non-standard instances of Numeric.  Unfortunately, this could result in the end wrapping around and hitting the beginning again, making the first element bad.

This patch fixes the behavior by altering the loop to end after the last element is encountered.  The particular flavor was chosen out of a few possibilities because it gave the best microbenchmarks on both large and small ranges.

Test written.  While testing, a bug was also uncovered in NumericRange, and was also fixed.  In brief, the logic around sum is rather complex since division is not unique when you have overflow.  Floating point has its own complexities, too.

Also updated incorrect test t4658 that insisted on incorrect answers (?!) and added logic to make sure it at least stays self-consistent, and fixed the range.scala test which used the same wrong (overflow-prone) formula that the Range collection did.