SI-9379  Added toString to .zipped to allow Stream etc to short-circuit
Tuple2Zipped and Tuple3Zipped would try to compute a hash code when .toString was called on them.  This overrides toString to print (collection1, collection2).zipped instead, using the collection's own toString method.  This allows collections that have a toString but not a hashCode (such as Iterator.from(0) and s = 1 #:: s) to print out as they usually do.

JUnit test to verify the deferral to collections' .toString.