Fix a long sleeper bug
When setting the denotation we did not change the checked period.
This can lead to a situation where a denotation is set to NoDenotation,
yet the checked period is something else. This means in effect the denotation
will vanish at the checked period. This bug caused the junit test failure
about "non-member exception" for sourceFile in DottyBackendInterface.