Remove ThreadPoolExecutor fallback in ExecutionContextImpl
The method createDefaultExecutorService had a fallback if the creation
of a ForkJoinPool didn't succeed. This was necessary, because Scala
shipped its own version of FJP, and the dependency on sun.misc.Unsafe
(which is not an "offical" official API) made portability slightly
questionable.

Now that we can assume that FJP is supplied by the JDK, this concern
goes away.