SI-9555 Refactor SystemProperties to avoid mutable map
A mutable map with help text was lazy populated if and when
lazy vals containing `BooleanProp` for a particular system was
initialized.

Reviewing this code, I noticed two problems: the class doesn't appear
to be designed for multithreaded access (e.g. someone could call
`help` while the map is being updated.), and the help method
won't return anything until the vals below are accessed for
the first time.

We don't use the help method anywhere else, but kept it around
in the unlikely event it was being used by some other tool.

I've done away with the mutable map by just moving the help messages
into a match in the help method itself.