Fix problems arising when hierarchies of classes are under completion
Fix problems arising when hierarchies of classes are under completion at the same time.
In this case it can happen that we see a subclass (e.g. Arrays.scala) which depends
on a superclass (e.g. GenTraversableLike.scala) that itself does not have its parents
defined yet. Previously, several things went wrong here

 - One of the base classes would be marked as frozen, even though it dod not have all
   members entered yet. This led to an error in finger printing.
 - The baseclasses and super class bits of the subclass would be computed before the parents
   of the middle class were known. The baseclasses would then be chached, leading to
   false results for isDerivedFrom.

We need to refine the logic for computing base classes, super class bits, and fingerprints
to account for that issue.