More refined treatement of pattern type variables
Previously all lower case names were treated as variables
in patterns. But that made code like

    x: cls

crash the compiler if `cls` was a class. Also, it owuld preventing
this idiom unless one wrote

    x: `cls`

We now do it like scalac and treat lower case names as variables
only in arguments of types.