Increase the maximum heap size on Jenkins
We're getting a lot of OutOfMemoryException when the maximum size is 1
GB, but we cannot increase it too much without using up all the memory
available on the Jenkins instances, let's see if 1.1 GB is enough.

Also stop using a custom -Xss, the default of 1 MB should be good enough.